general: lots of cleanup, compartmentalization
- removed the capability to run LevelEditor.py directly
- moved LevelEditorPanel to its own file
- moved VisGroupsEditor to its own file
- moved LEAvatar to its own file
- cleanup some more if statements
- cleanup some repetitive random.choices
